requirements.txt
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
pypdf>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pydub>=0.25.1
# Kokoro TTS (optional, only needed for --tts-engine kokoro)
torch>=2.0.0
kokoro>=0.9.2
soundfile>=0.12.1
numpy>=1.24.0
